Fixed: wflow-core - Form Row Delete Datalist Action - Error when no record is selected. (#801)
git-svn-id: http://dev.joget.org/svn/jw-community/trunk@430 1bff935d-2fe2-40b7-b00f-c2aca69a7612

+++ wflow-core/src/main/java/org/joget/apps/datalist/lib/FormRowDeleteDataListAction.java
@@ -65,15 +65,17 @@ public class FormRowDeleteDataListAction extends DataListActionDefault {
     public DataListActionResult executeAction(DataList dataList, String[] rowKeys) {
         DataListActionResult result = null;
 
-        String formDefId = getPropertyString("formDefId");
-        String tableName = getSelectedFormTableName(formDefId);
-        if (tableName != null) {
-            FormDataDao formDataDao = (FormDataDao) FormUtil.getApplicationContext().getBean("formDataDao");
-            formDataDao.delete(formDefId, tableName, rowKeys);
+        if (rowKeys != null && rowKeys.length > 0) {
+            String formDefId = getPropertyString("formDefId");
+            String tableName = getSelectedFormTableName(formDefId);
+            if (tableName != null) {
+                FormDataDao formDataDao = (FormDataDao) FormUtil.getApplicationContext().getBean("formDataDao");
+                formDataDao.delete(formDefId, tableName, rowKeys);
 
-            result = new DataListActionResult();
-            result.setType(DataListActionResult.TYPE_REDIRECT);
-            result.setUrl("REFERER");
+                result = new DataListActionResult();
+                result.setType(DataListActionResult.TYPE_REDIRECT);
+                result.setUrl("REFERER");
+            }
         }
 
         return result;
